fixed: floats and doubles where being incorrectly passed as longs.
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk@31643 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/lib/sqLite/mysqldataset.cpp
@@ -1308,7 +1308,7 @@ bool MysqlDataset::query(const char *query) {
       case MYSQL_TYPE_DOUBLE:
         if (row[i] != NULL)
         {
-          v.set_asDouble(atol(row[i]));
+          v.set_asDouble(atof(row[i]));
         }
         else
         {
